Converting PlayerIcon to functional component and introducing a condition to render player's country flag improves performance on this page by approx 30-40%% on my machine (1.1s on prod vs 0.7s locally).

However there is another performance issue: tooltips used abundantly. I tried to render this page without any tooltips and it further improved performance by about 40% (0.7s with tooltips vs 0.4s without). Surely we cannot remove tooltips here, but refactoring them in a way that does not have such a big impact on performance looks like a big overhaul which will influence all UI and not just this one page. If I have some spare time I'll probably take a shot at it.

0.7s and 0.4s still looks very unoptimized, so may be it is better not to close the corresponding issue and investigate other options for improvements