@SergeyKhval 2020-11-02
Yes, this is mergable. Just do not close corresponding issue